requirements.txt
langchain
unstructured
python-multipart
chromadb
tiktoken
openai
python-dotenv
fastapi
boto3
uvicorn